jobs/metadata-generator/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
langchain==0.3.13
langchain-openai==0.2.14
openai==1.54.5
